backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

# Real integrations
openai==1.30.0
httpx==0.27.0

# pgvector / Postgres (needed when DATABASE_URL is set)
asyncpg==0.29.0
pgvector==0.2.5
sqlalchemy[asyncio]==2.0.30
